Audiosignal is now captured and processed.

+++ app/src/main/java/com/exaltead/sceneclassifier/data_extraction/IAudioBufferer.kt
@@ -2,5 +2,5 @@ package com.exaltead.sceneclassifier.data_extraction
 
 interface IAudioBufferer {
 
-    fun takeShortAudioRecord(duration: Double): FloatArray
+    fun takeShortAudioRecord(duration: Double): ShortArray
 }
