+++ app/src/main/java/com/exaltead/sceneclassifier/data_extraction/WawBuffererI.kt
@@ -1,7 +1,7 @@
 package com.exaltead.sceneclassifier.data_extraction
 
 class WawBufferer : IAudioBufferer {
-    override fun takeShortAudioRecord(duration: Double): FloatArray {
+    override fun takeShortAudioRecord(duration: Double): ShortArray {
         TODO("not implemented")
     }
